refactor detailPage and EpisodeDetailPage
