+++ src/App.jsx
@@ -7,27 +7,38 @@ import DetailPage from './pages/DetailPage'
 import EpisodeDetailPage from './pages/EpisodeDetailPage'
 import ErrorPage from './pages/ErrorPage'
 import Header from './components/Header'
+import PodcastPage from './pages/PodcastPage';
 
 function App () {
-  const location = useLocation()
 
   return(
     <>
       <Header />
-      <Routes location={location}>
+      <Routes>
         <Route 
           path="/" 
           element={<MainPage />} 
         />
+        <Route
+          path='/podcast/:podcastId'
+          element={<PodcastPage />}
+          children={
+            <Route
+              path='/podcast/:podcastId/episode/:episodeId'
+              element={<EpisodeDetailPage />}
+            />
+          }
+        />
+        
         {/* DetailPage and EpisodeDetail could be in one component wiht conditional rendering as they share component */}
-        <Route 
+        {/* <Route 
           path="/podcast/:podcastId" 
           element={<DetailPage />}
-        />
-        <Route 
+        /> */}
+        {/* <Route 
           path="/podcast/:podcastId/episode/:episodeId" 
           element={<EpisodeDetailPage />}
-        />
+        /> */}
         {/* <Route 
           path="*" 
           element={<ErrorPage />}
