+++ src/components/EpisodeList.jsx
@@ -9,6 +9,7 @@ const EpisodeList = ({ podcastDetail }) => {
     const dispatch = useDispatch()
 
     const handleOnClickEpisode = (episode) => {
+        console.log(episode)
         // We add the episode audio to the global state to navigate to the episode detail page
         dispatch(setLoadingState(true))
 
@@ -38,11 +39,13 @@ const EpisodeList = ({ podcastDetail }) => {
                             key={`${episode.id}_${index}`} 
                             className={`flex flex-row justify-between p-2 ${index % 2 === 0 ? 'bg-gray-100' : 'bg-white'} `}
                         >
+                            <div className='flex w-8/12'>
                                 <Link 
                                     onClick={() =>handleOnClickEpisode(episode)}
                                     to={`/podcast/${podcastDetail.id}/episode/${episode.id}`}>
                                     <h3 className='text-start'>{episode.title}</h3>
                                 </Link>
+                            </div>
                             <div className="flex w-2/12">
                                 <p className='text-right'>{formatDateMMDDYYYY(episode.pubDate)}</p>
                             </div>
